Factor Activity out into separate types file
