+++ frontend/src/components/WorkweekHustle.tsx
@@ -2,7 +2,9 @@ import * as React from 'react';
 import _ from 'lodash'
 import { useQuery, gql } from '@apollo/client';
 
+import Activity from '../types/Activity';
 import UserLeaderboard from './UserLeaderboard';
+import UserActivityLog from './UserActivityLog';
 import {UserData} from './UserLeaderboard';
 
 export const FETCH_ACTIVITIES_QUERY = gql`
@@ -18,13 +20,29 @@ export const FETCH_ACTIVITIES_QUERY = gql`
     }
 `;
 
-type Activity = {
-    user: string;
-    createdAt: number;
-    recordDate: number;
-    steps: number;
-    activeMinutes: number;
-    distanceKm: number;
+export function getLatestActivityPerUserPerDay(activities: Activity[]): Activity[] {
+    return _.chain(activities)
+        .groupBy(
+            (activity: Activity) : string => {
+                return activity.user + "|" + activity.recordDate.toString();
+            }
+        )
+        .values()
+        .map((activities: Activity[]): Activity => {
+            return _.maxBy(activities, 'createdAt') || {
+                'user': 'unknown',
+                'createdAt': 0,
+                'recordDate': 0,
+                'steps': 0,
+                'activeMinutes': 0,
+                'distanceKm': 0,
+            }
+        })
+        .value();
+}
+
+export function getActivityLogs(activities: Activity[]): Activity[] {
+    return activities;
 }
 
 type WorkweekHustleProps = {
@@ -54,43 +72,31 @@ const WorkweekHustle = ({id, users, createdAt, startAt, endAt}: WorkweekHustlePr
     // There might be many logs for a single date.
     // Retrieve just the latest log for a given date.
     const activities: Activity[] = fetchActivities.data.activities;
-    const userData: UserData[] = _.chain(activities)
-        .groupBy(
-            (activity: Activity) : string => {
-                return activity.user + "|" + activity.recordDate.toString();
-            }
-        )
-        .values()
-        .map((activities: Activity[]): Activity => {
-            return _.maxBy(activities, 'createdAt') || {
-                'user': 'unknown',
-                'createdAt': 0,
-                'recordDate': 0,
-                'steps': 0,
-                'activeMinutes': 0,
-                'distanceKm': 0,
-            }
-        })
+    const leaderboardData: UserData[] = getLatestActivityPerUserPerDay(activities)
         .map((activity: Activity) => {
             return {
                 "name": activity.user,
                 "value": activity.steps,
                 "unit": "steps",
             }
-        })
-        .value();
+        });
+
+    const activitylogData: Activity[] = getActivityLogs(activities);
 
     return (
-        <UserLeaderboard
-            challengeName={"Workweek Hustle"}
-            id={id}
-            users={users}
-            userData={userData}
-            createdAt={createdAt}
-            startAt={startAt}
-            endAt={endAt}
-            unit={"steps"}
-        />
+        <div>
+            <UserLeaderboard
+                challengeName={"Workweek Hustle"}
+                id={id}
+                users={users}
+                userData={leaderboardData}
+                createdAt={createdAt}
+                startAt={startAt}
+                endAt={endAt}
+                unit={"steps"}
+            />
+            <UserActivityLog activities={activitylogData} />
+        </div>
     );
 };
