+++ frontend/src/types/Activity.tsx
@@ -1,0 +1,10 @@
+type Activity = {
+    user: string;
+    createdAt: number;
+    recordDate: number;
+    steps: number;
+    activeMinutes: number;
+    distanceKm: number;
+}
+
+export default Activity;
